+++ cx-portal/src/components/pages/AppReleaseProcess/components/AppMarketCard/index.tsx
@@ -29,10 +29,11 @@ import {
   Checkbox,
   PageNotifications,
   LogoGrayData,
+  SelectList,
 } from 'cx-portal-shared-components'
 import { useTranslation } from 'react-i18next'
 import { Grid, Divider, Box } from '@mui/material'
-import { useState } from 'react'
+import { useState, useEffect } from 'react'
 import KeyboardArrowLeftIcon from '@mui/icons-material/KeyboardArrowLeft'
 import HelpOutlineIcon from '@mui/icons-material/HelpOutline'
 import {
@@ -41,13 +42,20 @@ import {
   useAddCreateAppMutation,
   useUpdateDocumentUploadMutation,
   useCasesItem,
+  useFetchSalesManagerDataQuery,
+  salesManagerType,
+  useSaveAppMutation,
 } from 'features/appManagement/apiSlice'
 import { useNavigate } from 'react-router-dom'
 import { Controller, useForm } from 'react-hook-form'
 import { Dropzone } from 'components/shared/basic/Dropzone'
 import '../ReleaseProcessSteps.scss'
 import { useDispatch, useSelector } from 'react-redux'
-import { appStatusDataSelector, increment } from 'features/appManagement/slice'
+import {
+  appIdSelector,
+  appStatusDataSelector,
+  increment,
+} from 'features/appManagement/slice'
 import { setAppId } from 'features/appManagement/actions'
 import { isString } from 'lodash'
 import Patterns from 'types/Patterns'
@@ -64,7 +72,7 @@ type FormDataType = {
     leadPictureUri: File | null | string
     alt?: string
   }
-  salesManagerId: string
+  salesManagerId: string | null
 }
 
 export const ConnectorFormInputField = ({
@@ -181,14 +189,35 @@ export default function AppMarketCard() {
   const { t } = useTranslation()
   const navigate = useNavigate()
   const dispatch = useDispatch()
-
+  const appId = useSelector(appIdSelector)
   const [pageScrolled, setPageScrolled] = useState(false)
   const useCasesList = useFetchUseCasesQuery().data || []
   const appLanguagesList = useFetchAppLanguagesQuery().data || []
   const [addCreateApp] = useAddCreateAppMutation()
+  const [saveApp] = useSaveAppMutation()
   const [updateDocumentUpload] = useUpdateDocumentUploadMutation()
   const [appCardNotification, setAppCardNotification] = useState(false)
   const appStatusData = useSelector(appStatusDataSelector)
+  const salesManagerList = useFetchSalesManagerDataQuery().data || []
+  const defaultSalesManagerValue = {
+    userId: null,
+    firstName: 'none',
+    lastName: '',
+    fullName: 'none',
+  }
+  const [salesManagerListData, setSalesManagerListData] = useState<
+    salesManagerType[]
+  >([defaultSalesManagerValue])
+  const [salesManagerId, setSalesManagerId] = useState(null)
+
+  useEffect(() => {
+    if (salesManagerList.length > 0) {
+      let data = salesManagerList?.map((item) => {
+        return { ...item, fullName: `${item.firstName} ${item.lastName}` }
+      })
+      setSalesManagerListData(salesManagerListData.concat(data))
+    }
+  }, [salesManagerList])
 
   const defaultValues = {
     title: appStatusData?.title,
@@ -197,7 +226,7 @@ export default function AppMarketCard() {
     useCaseCategory: appStatusData?.useCase,
     appLanguage: appStatusData?.supportedLanguageCodes,
     //To do: to be changed once api is available
-    salesManagerId: 'ac1cf001-7fbc-1f2f-817f-bce058020001',
+    salesManagerId: salesManagerId,
     shortDescriptionEN:
       appStatusData?.descriptions?.filter(
         (appStatus: any) => appStatus.languageCode === 'en'
@@ -263,7 +292,7 @@ export default function AppMarketCard() {
         data.uploadImage.leadPictureUri !== null &&
         Object.keys(data.uploadImage.leadPictureUri).length > 0 &&
         Object.values(data.uploadImage.leadPictureUri)[0],
-      salesManagerId: data.salesManagerId,
+      salesManagerId: salesManagerId,
       useCaseIds: data.useCaseCategory,
       descriptions: [
         {
@@ -282,18 +311,35 @@ export default function AppMarketCard() {
     }
 
     const uploadImageValue = getValues().uploadImage.leadPictureUri
-    await addCreateApp(saveData)
-      .unwrap()
-      .then((result) => {
-        if (isString(result)) {
-          uploadDocumentApi(result, 'APP_LEADIMAGE', uploadImageValue)
-          dispatch(setAppId(result))
+
+    if (appId) {
+      const saveAppData = {
+        appId: appId,
+        body: saveData,
+      }
+
+      await saveApp(saveAppData)
+        .unwrap()
+        .then(() => {
           dispatch(increment())
-        }
-      })
-      .catch((error: any) => {
-        setAppCardNotification(true)
-      })
+        })
+        .catch(() => {
+          setAppCardNotification(true)
+        })
+    } else {
+      await addCreateApp(saveData)
+        .unwrap()
+        .then((result) => {
+          if (isString(result)) {
+            uploadDocumentApi(result, 'APP_LEADIMAGE', uploadImageValue)
+            dispatch(setAppId(result))
+            dispatch(increment())
+          }
+        })
+        .catch(() => {
+          setAppCardNotification(true)
+        })
+    }
   }
 
   const uploadDocumentApi = async (
@@ -454,8 +500,8 @@ export default function AppMarketCard() {
 
             <div className="form-field">
               {['shortDescriptionEN', 'shortDescriptionDE'].map(
-                (item: string) => (
-                  <>
+                (item: string, i) => (
+                  <div key={i}>
                     <ConnectorFormInputField
                       {...{
                         control,
@@ -523,7 +569,7 @@ export default function AppMarketCard() {
                         ? getValues().shortDescriptionEN.length
                         : getValues().shortDescriptionDE.length) + `/255`}
                     </Typography>
-                  </>
+                  </div>
                 )
               )}
             </div>
@@ -610,6 +656,24 @@ export default function AppMarketCard() {
                   },
                 }}
               />
+            </div>
+
+            <div className="form-field">
+              <SelectList
+                defaultValue={defaultSalesManagerValue}
+                items={salesManagerListData}
+                label={t('content.apprelease.appMarketCard.salesManager')}
+                placeholder={t(
+                  'content.apprelease.appMarketCard.salesManagerPlaceholder'
+                )}
+                onChangeItem={(e) => {
+                  setSalesManagerId(e.userId)
+                }}
+                keyTitle={'fullName'}
+              />
+              <Typography variant="body2">
+                {t('content.apprelease.appMarketCard.salesManagerDescription')}
+              </Typography>
             </div>
 
             <div className="form-field">