+++ cx-portal/src/components/pages/AppReleaseProcess/components/AppPage/index.tsx
@@ -237,8 +237,8 @@ export default function AppPage() {
       </Grid>
       <form className="header-description">
         <div className="form-field">
-          {['longDescriptionEN', 'longDescriptionDE'].map((item: string) => (
-            <>
+          {['longDescriptionEN', 'longDescriptionDE'].map((item: string, i) => (
+            <div key={i}>
               <ConnectorFormInputField
                 {...{
                   control,
@@ -300,7 +300,7 @@ export default function AppPage() {
                   : getValues().longDescriptionDE.length) +
                   `/${longDescriptionMaxLength}`}
               </Typography>
-            </>
+            </div>
           ))}
         </div>
 
@@ -347,8 +347,8 @@ export default function AppPage() {
           'uploadTechnicalGuide',
           'uploadDataContract',
           'uploadAppContract',
-        ].map((item: string) => (
-          <>
+        ].map((item: string, i) => (
+          <div key={i}>
             <Divider sx={{ mb: 2, mr: -2, ml: -2 }} />
             <div className="form-field" key={item}>
               <InputLabel sx={{ mb: 3, mt: 3 }}>
@@ -406,7 +406,7 @@ export default function AppPage() {
                   </Typography>
                 )}
             </div>
-          </>
+          </div>
         ))}
 
         <Divider sx={{ mb: 2, mr: -2, ml: -2 }} />